dspy_service/requirements.txt
# CourseLLM DSPy Service Dependencies
# Python 3.11 compatible

# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# DSPy for prompt engineering (v3+)
dspy-ai>=3.0.4

# Google AI / Gemini
google-generativeai==0.3.2

# Environment management
python-dotenv==1.0.0

# JSON handling
orjson==3.9.10

# Optional: JSON repair for handling malformed JSON from LLM responses
# If not installed, the service will still work but may fail on malformed JSON
# json-repair>=0.23.0

